backend/requirements-ci.txt
Django>=6.0,<6.1
python-decouple>=3.8
pillow>=10.0
openpyxl>=3.1

djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
django-axes>=7.0
django-storages>=1.14

pytest>=8.0
pytest-django>=4.8
pytest-cov>=5.0
pytest-xdist>=3.6
pytest-sugar>=1.0
